tell now that I reluctantly maintain this package and up the version
The package was removed ("archived") from CRAN because it did not
pass CRAN tests. I have fixed those issues in this package.
The fixes have been communicated to the GitHub repository of this
package as an issue https://github.com/jwindle/BayesLogit/issues/13
and a pull request https://github.com/jwindle/BayesLogit/pull/14,
with no reaction. We need this package, and therefore I edited the
DESCRIPTION as if this is a new version and I am maintainer. I do
this reluctantly, and I hope the real developers will take over with
the package.